backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.0",
    "fastapi[standard]>=0.116.1",
    "motor>=3.7.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
